web/icon: move demo components to their own module, refine styles.
