+++ ui/web/src/app/common/colors.scss
