+++ ui/web/src/app/demo/styles/styles.component.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import 'src/app/common/base';
+@import 'src/app/common/colors';
 @import 'node_modules/sass-material-colors/sass/sass-material-colors-classes';
 
 section {
